+++ prp/parse/phenotype/tbprofiler.py
@@ -7,7 +7,6 @@ from ...models.phenotype import ElementTypeResult
 from ...models.phenotype import PredictionSoftware as Software
 from ...models.phenotype import ResistanceVariant
 from ...models.sample import MethodIndex
-from .utils import _default_variant
 
 LOG = logging.getLogger(__name__)
 
@@ -50,14 +49,16 @@ def _parse_tbprofiler_amr_variants(tbprofiler_result) -> Tuple[ResistanceVariant
 
     for hit in tbprofiler_result["dr_variants"]:
         var_type = "substitution"
+
         variant = ResistanceVariant(
             variant_type=var_type,
-            genes=[hit["gene"]],
-            phenotypes=hit["gene_associated_drugs"],
+            gene_symbol=hit["gene"],
+            phenotypes=[],
             position=int(hit["genome_pos"]),
             ref_nt=hit["ref"],
             alt_nt=hit["alt"],
             depth=hit["depth"],
+            freq=float(hit["freq"]),
             ref_database=tbprofiler_result["db_version"]["name"],
             type=hit["type"],
             nucleotide_change=hit["nucleotide_change"],
@@ -68,7 +69,7 @@ def _parse_tbprofiler_amr_variants(tbprofiler_result) -> Tuple[ResistanceVariant
         results.append(variant)
 
     if not results:
-        results = _default_variant().mutations
+        results = [ResistanceVariant()]
         return results
 
     return results