+++ tests/fixtures/__init__.py
@@ -2,3 +2,4 @@
 
 from .ecoli import *
 from .saureus import *
+from .mtuberculosis import *